refactor: split generate.go, decompose Scaffold.Run, extract hooks package
- Split commands/generate.go (544 lines) into 3 focused files:
  generate.go (command + execution), generate_list.go (discovery),
  generate_resolve.go (path resolution)
- Decompose Scaffold.Run() into 5 phase methods via runContext struct:
  loadConfig → confirmHooks → collectVars → planOutput → executeScaffold
- Extract hooks into standalone internal/hooks/ package (4 source + 2 test files)
  with minimal Confirmer interface to avoid scaffold dependency
- Move all hook unit tests to hooks/, keep only integration tests in scaffold/
