+++ internal/commands/generate_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/kaikenlabs/tag/internal/config"
 	"github.com/kaikenlabs/tag/internal/engine"
-	"github.com/kaikenlabs/tag/internal/scaffold"
+	"github.com/kaikenlabs/tag/internal/hooks"
 	"github.com/kaikenlabs/tag/internal/template"
 	"github.com/kaikenlabs/tag/internal/types/flags"
 	"github.com/kaikenlabs/tag/pkg/app"
@@ -379,12 +379,12 @@ Hello {{ .Name }}`
 }
 
 func TestUT_RunHooks_EmptyHooks(t *testing.T) {
-	err := runHooks([][]string{}, scaffold.HookPhasePreGen)
+	err := runHooks([][]string{}, hooks.HookPhasePreGen)
 	require.NoError(t, err)
 }
 
 func TestUT_RunHooks_NilHooks(t *testing.T) {
-	err := runHooks(nil, scaffold.HookPhasePreGen)
+	err := runHooks(nil, hooks.HookPhasePreGen)
 	require.NoError(t, err)
 }
 
@@ -404,7 +404,7 @@ func TestUT_RunHooks_ValidHook(t *testing.T) {
 	t.Cleanup(func() { _ = os.Chdir(oldDir) })
 
 	// Use relative path with ./ prefix to indicate it's a local file
-	err = runHooks([][]string{{"./test-hook.sh"}}, scaffold.HookPhasePreGen)
+	err = runHooks([][]string{{"./test-hook.sh"}}, hooks.HookPhasePreGen)
 	require.NoError(t, err)
 }
 
@@ -423,7 +423,7 @@ func TestUT_RunHooks_FailingHook(t *testing.T) {
 	t.Cleanup(func() { _ = os.Chdir(oldDir) })
 
 	// Use relative path with ./ prefix to indicate it's a local file
-	err = runHooks([][]string{{"./failing-hook.sh"}}, scaffold.HookPhasePreGen)
+	err = runHooks([][]string{{"./failing-hook.sh"}}, hooks.HookPhasePreGen)
 	require.Error(t, err)
 	assert.Contains(t, err.Error(), "hook failed")
 }
@@ -449,7 +449,7 @@ func TestUT_RunHooks_StopsOnFailure(t *testing.T) {
 	t.Cleanup(func() { _ = os.Chdir(oldDir) })
 
 	// Use relative paths with ./ prefix to indicate they are local files
-	err = runHooks([][]string{{"./hook1.sh"}, {"./hook2.sh"}}, scaffold.HookPhasePreGen)
+	err = runHooks([][]string{{"./hook1.sh"}, {"./hook2.sh"}}, hooks.HookPhasePreGen)
 	require.Error(t, err)
 
 	// Verify second hook didn't run
@@ -466,13 +466,13 @@ func TestUT_RunHooks_NonexistentCommand(t *testing.T) {
 	require.NoError(t, err)
 	t.Cleanup(func() { _ = os.Chdir(oldDir) })
 
-	err = runHooks([][]string{{"nonexistent-command-xyz"}}, scaffold.HookPhasePreGen)
+	err = runHooks([][]string{{"nonexistent-command-xyz"}}, hooks.HookPhasePreGen)
 	require.Error(t, err)
 }
 
 func TestUT_RunHooks_PathCommand(t *testing.T) {
 	// Test that PATH commands work (e.g., "echo" should be found in PATH)
-	err := runHooks([][]string{{"echo", "hello"}}, scaffold.HookPhasePreGen)
+	err := runHooks([][]string{{"echo", "hello"}}, hooks.HookPhasePreGen)
 	require.NoError(t, err)
 }
 
@@ -491,7 +491,7 @@ func TestUT_RunHooks_RelativePathCommand(t *testing.T) {
 	t.Cleanup(func() { _ = os.Chdir(oldDir) })
 
 	// Relative path should resolve from working directory
-	err = runHooks([][]string{{"./myscript.sh"}}, scaffold.HookPhasePreGen)
+	err = runHooks([][]string{{"./myscript.sh"}}, hooks.HookPhasePreGen)
 	require.NoError(t, err)
 }
