+++ internal/hooks/interpreter.go
@@ -1,4 +1,4 @@
-package scaffold
+package hooks
 
 import (
 	"errors"
