+++ internal/hooks/interpreter_test.go
@@ -1,4 +1,4 @@
-package scaffold
+package hooks
 
 import (
 	"os"
